refactor: KanbanView.tsx 컴포넌트 분리 (CM-52)
